Working Undo/Redo Function
When an event is created/modified/deleted the action is logged in the db collection named EventActionLog with the action type, event data before and after action, the project the event is associated with, time when action is preformed, and if the action was undone.
The event will store a total of 5 undo/redo documents in the EventActionLog collection. Any additional event actions will delete the oldest action from the EventActionLog Collection.
Undo will revert the most recent change made, changing the assocated document is_undone flag to be true.
Redo will revert the most recent undone action that has the is_undone flag set to true.